Kelsi Worrell
Kelsi exploded on to the national and international scene during her
junior season at the University of Louisville. She was the NCAA Champion
in the 100-fly with the time of 49.81, which set an American Record and
an NCAA Record. She then took the NCAA champion in the 200 butterfly
with the time of 1:51.11. She was third at NCAAs in 50-Free with a 21.62
and was part of the second place 200-medley relay, swimming the fly leg
in 21.96.

The Cards were, 6th in 400 Medley Relay AT NCAAs thanks to her third leg
split of 50.09. At NCAAs, she helped the Cards to an 11th in 400-free
relay, swimming the second leg in 47.66. At the ACC Championships, she won
100-free with a ACC meet record 47.71 and then won 100-fly win 50.17.
She was second in 50 free, posting a 22.12 and swam the third leg of 200
Medley relay (22.34) to win ACC. She swam the lead off leg of the
200-free relay in 22.28 and helped the Cards finish fifth at ACCs. She
was the second leg of the ACC bronze medal 400 Free Relay,
splitting 47.39.

She was named ACC Swimmer of the Year.  In her first international
competition, she won Pan Am Gold in the 100-fly going 57.24.  She helped
Team USA win Pan Am Gold in the 4 x 100 Medley Relay going 3:57.35 with
a split 57.63 which was not only a Pan Am record but became the first women
to break four minutes. She won Pan Am Silver in the 4 X 100 Free Relay with
a time of 3:37.28 with her splitting 54.03. At the Arena Pro Series in Mesa,
she was first in 400 Medley Relay with a split of 1:00.05.  She also won
the Arena Pro Series gold in the 100-M Fly with a time of 58.24.  She won a
gold medal at USA Nationals swimming the lead off leg of the championship
4x100 Free Relay in 54.73.  She also won gold at USA Nationals in the
100-M Fly with a time of 57.27 and was fourth at USA Nationals in 100 back
(1:00.37). She was named to the USA National Team.  She was a second team
CoSIDA Academic All-American.

http://www.gocards.com/roster.aspx?rp_id=5386